# comments/api/openapiv2/comments.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Example project: Comments API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Comments",
      "description": "Comments service"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/comment/create": {
      "post": {
        "operationId": "Comments_CreateComment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateCommentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateCommentRequest"
            }
          }
        ],
        "tags": [
          "Comments"
        ]
      }
    },
    "/comment/list": {
      "get": {
        "operationId": "Comments_GetComments",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetCommentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "productID",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Comments"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Comment": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "int64"
        },
        "userID": {
          "type": "string",
          "format": "int64"
        },
        "text": {
          "type": "string"
        },
        "ts": {
          "type": "string",
          "format": "date-time"
        }
      },
      "description": "Comment item",
      "title": "Comment",
      "required": [
        "ID",
        "userID",
        "productID",
        "text"
      ]
    },
    "v1CreateCommentRequest": {
      "type": "object",
      "properties": {
        "userID": {
          "type": "string",
          "format": "int64"
        },
        "productID": {
          "type": "string",
          "format": "int64"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "v1CreateCommentResponse": {
      "type": "object",
      "properties": {
        "commentID": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1GetCommentsResponse": {
      "type": "object",
      "properties": {
        "productID": {
          "type": "string",
          "format": "int64"
        },
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Comment"
          }
        }
      }
    }
  }
}